requirements.txt
requests==2.22.0
python-dotenv==0.10.3
pytz==2019.3

azure-storage-blob
